PatternsViewModel splitted on PatternsViewModel and PatternViewModel. Correct update of items in ListBox.
