+++ Sources/PT.PM.PatternEditor/PatternUserControl.paml
@@ -1,9 +1,17 @@
-﻿<UserControl xmlns="https://github.com/avaloniaui" >
+﻿<UserControl xmlns="https://github.com/avaloniaui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:c="clr-namespace:PT.PM.PatternEditor;assembly=PT.PM.PatternEditor">
   <UserControl.Styles>
     <Style Selector="TextBlock">
       <Setter Property="VerticalAlignment" Value="Center"/>
     </Style>
   </UserControl.Styles>
+
+  <UserControl.DataTemplates>
+    <DataTemplate DataType="c:PatternViewModel" >
+      <TextBlock Text="{Binding Title}" />
+    </DataTemplate>
+  </UserControl.DataTemplates>
   
   <StackPanel Grid.Column="0" Gap="7">
     <Grid ColumnDefinitions="100,100,*" Height="30">
@@ -16,7 +24,8 @@
       </TextBlock>
     </Grid>
     
-    <ListBox Name="PatternsListBox" MinHeight="150" MaxHeight="150" Items="{Binding Patterns}" SelectedItem="{Binding SelectedPattern, Mode=OneWay}" AutoScrollToSelectedItem="true" />
+    <ListBox Name="PatternsListBox" MinHeight="150" MaxHeight="150" VirtualizationMode="Simple" Items="{Binding Patterns}" SelectedItem="{Binding SelectedPattern}" AutoScrollToSelectedItem="true" />
+    
     <Grid ColumnDefinitions="*,*,*,*">
       <Button Grid.Column="0" Content="Create" Command="{Binding CreatePattern}" Margin="0,0,7,0" />
       <Button Grid.Column="1" Content="Remove" Command="{Binding RemovePattern}" Margin="0,0,7,0" />
@@ -49,6 +58,7 @@
         <CheckBox Grid.Row="2" Grid.Column="1" IsChecked="{Binding IsHtmlLanguage}">Html</CheckBox>
       </Grid>
     </Grid>
+    
     <TextBlock Text="Description" />
     <TextBox AcceptsReturn="true" Text="{Binding Description}" Height="100" TextWrapping="Wrap" />
     <Grid ColumnDefinitions="100,*">
